+++ tests/testthat/test_daycent.R
@@ -21,28 +21,56 @@ env_daily<-function(out, yearIni = 2000) {
   return(df)
 }
 
-prd_daily<-function(out, yearIni = 2000, species = "") {
+prd_daily_new<-function(out, rcentury_tree, yearIni = 2000) {
+  decid <- rcentury_tree$df[rcentury_tree$df[,2]=="DECID",1]
+  species <- rcentury_tree$label
   yearMax <- yearIni + max(out$time) - 1
   dateIni <- as.Date(paste0(yearIni, "-01-01"))
   dateFin <- as.Date(paste0(yearMax, "-12-31"))
   days <- seq(dateIni, dateFin, by="day")
   months <- cut(days, breaks="months")
   unique_months <- unique(months)
+  leafdr <- c(rcentury_tree$df[rcentury_tree$df[,2]=="LEAFDR(1)",1],
+              rcentury_tree$df[rcentury_tree$df[,2]=="LEAFDR(2)",1],
+              rcentury_tree$df[rcentury_tree$df[,2]=="LEAFDR(3)",1],
+              rcentury_tree$df[rcentury_tree$df[,2]=="LEAFDR(4)",1],
+              rcentury_tree$df[rcentury_tree$df[,2]=="LEAFDR(5)",1],
+              rcentury_tree$df[rcentury_tree$df[,2]=="LEAFDR(6)",1],
+              rcentury_tree$df[rcentury_tree$df[,2]=="LEAFDR(7)",1],
+              rcentury_tree$df[rcentury_tree$df[,2]=="LEAFDR(8)",1],
+              rcentury_tree$df[rcentury_tree$df[,2]=="LEAFDR(9)",1],
+              rcentury_tree$df[rcentury_tree$df[,2]=="LEAFDR(10)",1],
+              rcentury_tree$df[rcentury_tree$df[,2]=="LEAFDR(11)",1],
+              rcentury_tree$df[rcentury_tree$df[,2]=="LEAFDR(12)",1])
+  wooddr1 <- rcentury_tree$df[rcentury_tree$df[,2]=="WOODDR(1)",1] #fraction of forest which is deciduous; the fraction of leaves which fall during senescence month or at the end of the growing season, used when decid = 1 or 2
+  wooddr2 <- rcentury_tree$df[rcentury_tree$df[,2]=="WOODDR(2)",1] #monthly death rate fraction for fine root component
+  wooddr3 <- rcentury_tree$df[rcentury_tree$df[,2]=="WOODDR(3)",1] #monthly death rate fraction for fine branch component
+  wooddr4 <- rcentury_tree$df[rcentury_tree$df[,2]=="WOODDR(4)",1] #monthly death rate fraction for large wood component
+  wooddr5 <- rcentury_tree$df[rcentury_tree$df[,2]=="WOODDR(5)",1] #monthly death rate fraction for coarse root component
   # Trim out to the complete years
   out <- out[1:length(unique_months), , drop = FALSE]
-  # fbrprd -> production of fine branches (gC/m2/yr)
-  # crtprd -> production of coarse roots (gC/m2/yr)
-  # rlwprd -> large wood component C production (gC/m2/yr)
-  # rlvprd -> leaves component C production (gC/m2/yr)
+  # fbrchc -> C in forest system fine branch component (g m-2).
+  # crootc -> C in forest system coarse root component (g m-2).
+  # rlwodc -> C in forest system large wood component (g m-2).
+  # frootc -> C in forest system fine root component (g m-2).
+  # rleavc -> C in forest system leaf component (g m-2).
+  # print(wooddr1)
   df <- data.frame(date = days, Step = 1:length(days), Species = species, Leaves = NA, Twigs = NA, SmallBranches = NA, LargeWood = NA, CoarseRoots = NA, FineRoots = NA)
   for(i in 1:length(unique_months)) {
     sel_i <- months == unique_months[i]
-    df$Leaves[sel_i] <- out$rlvprd[i]/365
+    nmonth_days <- sum(sel_i)
+    month_number <- as.numeric(format(days[sel_i], "%m")[1])
+    if(decid == 0) {
+      df$Leaves[sel_i] <- out$rleavc[i]*leafdr[month_number]/nmonth_days
+    } else {
+      if((i %% 12)==10) df$Leaves[sel_i] <- out$rleavc[i]*wooddr1/nmonth_days
+      else df$Leaves[sel_i] <- 0
+    }
     df$Twigs[sel_i] <- 0
-    df$SmallBranches[sel_i] <- out$fbrprd[i]/365
-    df$LargeWood[sel_i] <- out$rlwprd[i]/365
-    df$CoarseRoots[sel_i] <- out$crtprd[i]/365
-    df$FineRoots[sel_i] <- out$frtprd[i]/365
+    df$SmallBranches[sel_i] <- out$fbrchc[i]*wooddr3/nmonth_days
+    df$LargeWood[sel_i] <- out$rlwodc[i]*wooddr4/nmonth_days
+    df$CoarseRoots[sel_i] <- out$crootc[i]*wooddr5/nmonth_days
+    df$FineRoots[sel_i] <- out$frootc[i]*wooddr2/nmonth_days
   }
   return(df)
 }
@@ -88,13 +116,13 @@ test_that("DAYCENT can be run", {
   skip_on_ci()
   skip_on_cran()
   
-  load("~/OneDrive/mcaceres_work/model_development/medfate_evaluation/RcenturyBenchmark/century.RData")
-  out_ForestC <- readRDS("~/OneDrive/mcaceres_work/model_development/medfate_evaluation/RcenturyBenchmark/out_ForestC.rds")
+  load("/home/miquel/OneDrive/mcaceres_work/model_development/medfate_evaluation/RcenturyBenchmark/Rdata/RC2m_Process.RData")
+  duke_tree <- tree$`3`
   
-  environmentalConditions <- env_daily(out_WaterTemp)
-  litterProduction <- prd_daily(out_ForestC)
+  environmentalConditions <- env_daily(out_duke)
+  litterProduction <- prd_daily_new(out_duke, duke_tree)
   
-  paramsLitterDecomposition <- sp_params(tree$`1`)
+  paramsLitterDecomposition <- sp_params(duke_tree)
   paramsAnatomy <- paramsLitterDecomposition[,c("Species"), drop = FALSE]
   paramsAnatomy$WoodDensity <- 0.5
   litterProduction$Species <- paramsLitterDecomposition$Species[1]