+++ NEWS.md
@@ -1,4 +1,5 @@
 # medfate 5.0.0
+* Partial rhizosphere overlap set to default
 * Internal restructuration to avoid Rcpp structures (NumericVector, NumericMatrix, DataFrame,...)
 * Approximate speed increase (Granier): spwb_day()/growth_day() x 4.5, spwb(1 yr) x 8.5, growth(1 yr) x 8.5
 * Approximate speed increase (Sperry): spwb_day()/growth_day() x 2.5, spwb(1 yr) x 4.0, growth (1 yr) x 1.5
@@ -6,6 +7,7 @@
 * Rcpp modules "runners" to run single-site, multiple-site and watershed simulations while avoiding unnecessary input/output.
 * Default values for leaf/stem cavitation recovery set to "rate" to avoid anomalous behavior observed with "annual"
 * Leaf defoliation for "spwb" simulations made coherent with defoliation in "growth". Control parameter "cavitationInducedDefoliation" added.
+
 
 # medfate 4.9.0
 * Forest objects can now have a seedling bank, snag data and litter data